Modified models for shopping Cart

+++ src/main/java/com/adminportal/domain/PromoCodes.java
@@ -23,7 +23,7 @@ public class PromoCodes {
 	
 	//Either Specific User or All
 	private String promoType;
-	private int promoValue;
+	private BigDecimal promoValue;
 	
 	private boolean promoStatus;
 
@@ -99,11 +99,11 @@ public class PromoCodes {
 		this.promoType = promoType;
 	}
 
-	public int getPromoValue() {
+	public BigDecimal getPromoValue() {
 		return promoValue;
 	}
 
-	public void setPromoValue(int promoValue) {
+	public void setPromoValue(BigDecimal promoValue) {
 		this.promoValue = promoValue;
 	}
